requirements.txt
google-cloud-vision
google-generativeai
streamlit